new features added-edit left

+++ src/App.jsx
@@ -5,6 +5,7 @@ import NavBar from "./components/NavBar";
 import Card from "./components/Card";
 import BlogDetail from "./components/BlogDetail";
 import CoverPage from "./components/CoverPage";
+import Swal from "sweetalert2";
 
 function App() {
   const [blogs, setBlogs] = useState([]);
@@ -16,9 +17,26 @@ function App() {
   }, []);
 
   const handleDelete = (blogToDelete) => {
-    const updatedBlogs = blogs.filter(blog => blog !== blogToDelete);
-    setBlogs(updatedBlogs);
-    localStorage.setItem("blogs", JSON.stringify(updatedBlogs));
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!"
+    }).then((result) => {
+      if (result.isConfirmed) {
+        const updatedBlogs = blogs.filter(blog => blog !== blogToDelete);
+        setBlogs(updatedBlogs);
+        localStorage.setItem("blogs", JSON.stringify(updatedBlogs));
+        Swal.fire({
+          title: "Deleted!",
+          text: "Your blog has been deleted.",
+          icon: "success"
+        });
+      }
+    });
   };
 
   const handleLike = (index) => {
